Add internet permission. Move Card class to data directory. Move CardDisplay activity and related resources to card.
