+++ app/src/main/AndroidManifest.xml
@@ -2,6 +2,8 @@
 <manifest package="com.bizdesc.birhanu.myapplication"
           xmlns:android="http://schemas.android.com/apk/res/android">
 
+  <uses-permission android:name="android.permission.INTERNET" />
+
   <application
     android:allowBackup="true"
     android:icon="@mipmap/ic_launcher"
@@ -27,7 +29,7 @@
       android:value="@integer/google_play_services_version"/>
 
     <activity
-      android:name=".DisplayCardActivity"
+      android:name=".CardActivity"
       android:label="@string/title_activity_display_message"
       android:parentActivityName=".LoginActivity"
       android:theme="@style/AppTheme.NoActionBar">
