+++ app/src/main/res/layout/activity_card.xml
@@ -6,7 +6,7 @@
   android:layout_width="match_parent"
   android:layout_height="match_parent"
   android:fitsSystemWindows="true"
-  tools:context=".DisplayCardActivity">
+  tools:context=".CardActivity">
 
   <android.support.design.widget.AppBarLayout
     android:layout_width="match_parent"
@@ -22,7 +22,7 @@
 
   </android.support.design.widget.AppBarLayout>
 
-  <include layout="@layout/content_display_card"/>
+  <include layout="@layout/content_card"/>
 
   <android.support.design.widget.FloatingActionButton
     android:id="@+id/fab"
