+++ app/src/main/res/layout/content_card.xml
@@ -10,7 +10,7 @@
   android:paddingRight="@dimen/activity_horizontal_margin"
   android:paddingTop="@dimen/activity_vertical_margin"
   app:layout_behavior="@string/appbar_scrolling_view_behavior"
-  tools:context=".DisplayCardActivity"
-  tools:showIn="@layout/activity_display_card">
+  tools:context=".CardActivity"
+  tools:showIn="@layout/activity_card">
 
 </RelativeLayout>
